Adjusted Dashboard and Analytics
