+++ src/components/analytics/AuthorCard.tsx
@@ -1,6 +1,6 @@
 "use client";
 import Image from "next/image";
-
+import Image2 from "@/assets/Images/footerimage.png";
 type AuthorCardProps = {
   name: string;
   image: string;
@@ -9,15 +9,11 @@ type AuthorCardProps = {
 const AuthorCard: React.FC<AuthorCardProps> = ({ name, image }) => (
   <div className="text-center">
     <Image
-      src={image}
+      src={Image2}
       alt={name}
       width={100}
       height={100}
       className="rounded-full object-cover mx-auto"
-      onError={(e) => {
-        (e.target as HTMLImageElement).src =
-          "https://placehold.co/100x100/CCCCCC/000000?text=Author";
-      }}
     />
     <p className="mt-2 font-semibold text-gray-800">{name}</p>
   </div>
