+++ src/components/analytics/Icon.tsx
@@ -1,4 +1,4 @@
-"use client"
+"use client";
 import React from "react";
 
 type IconName =
